apps/scanner-engine/requirements.txt
kafka-python
requests
beautifulsoup4
playwright
